refactor: contract supports to zoe
